added health system
I have succesfully added a simple health system to the player for when it collides with an enemy it will take damage. If the variable being used is equal or less than 10 then it will change the game to an empty room.

Fixed bug where player and enemy colide it would quit the game if there was a health system in place.

current issue of not showing death screen when it should.